requirements.txt
pytesseract
pdf2image
nltk
natsort
matplotlib
seaborn
Pillow
camelot-py[cv]
pandas
transformers
torch
flask
pillow
flake8
pytest